requirements.txt
apache-airflow==2.7.3
pandas==2.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
SQLAlchemy==1.4.50
virtualenv==20.24.6